Render background correctly. Simplify convergence. Check the web to see if Obama won.
Use 2 spaces.
Handle the case when we are on the day of the election or after.
Check openkeyval for whether or not Obama won. (Assume he did =)
Use colorWithCalibratedRed to set the color properly.
Also, instead of relative RGB values, use generics to get more accurate
values.
Wait both color states, and do it at the beginning of the transition.